refactor(hooks): export useAllowClear from hooks entry
