+++ components/date-picker/generatePicker/generateSinglePicker.tsx
@@ -8,8 +8,7 @@ import { merge } from '@rc-component/util';
 import { clsx } from 'clsx';
 
 import ContextIsolator from '../../_util/ContextIsolator';
-import { useZIndex } from '../../_util/hooks';
-import useAllowClear from '../../_util/hooks/useAllowClear';
+import { useAllowClear, useZIndex } from '../../_util/hooks';
 import { getMergedStatus, getStatusClassNames } from '../../_util/statusUtils';
 import type { AnyObject } from '../../_util/type';
 import { devUseWarning } from '../../_util/warning';
@@ -183,6 +182,7 @@ const generatePicker = <DateType extends AnyObject = AnyObject>(
 
       // ===================== Icon =====================
       const [, removeIcon] = useIcons(props, prefixCls);
+
       const mergedAllowClear = useAllowClear({
         componentName: displayName,
         allowClear,
